refactor: isolate import map
